Atualização Migrates

+++ db/migrate/20190505121445_create_questionaries.rb
@@ -1,12 +1,12 @@
 class CreateQuestionaries < ActiveRecord::Migration[5.2]
   def change
     create_table :questionaries do |t|
-      t.integer :question1a
-      t.integer :question1b
-      t.integer :question2a
-      t.integer :question2b
-      t.integer :question3a
-      t.integer :question3b
+      t.decimal :question1a
+      t.decimal :question1b
+      t.decimal :question2a
+      t.decimal :question2b
+      t.decimal :question3a
+      t.decimal :question3b
       t.string :question4a
       t.string :question4b
       t.date :created_at
